fix: address reviewer comments + update docs
- Server auto-triggers migration on first start (non-fatal try/catch)
- config set validates numeric ranges (chunkSize 64-4096, tokenMax 100-16384)
- deleteConfigFile only catches ENOENT, rethrows permission errors
- Remove config cache mechanism (unnecessary, each process reads once)
- Move MEMORY.md boost before maxResults truncation
- Extract migrateFromLegacyDirs to config.ts (shared by server + setup)
- README: update model config reference, add CLI override docs, add
  Upgrading section, document MEMORY.md boost in How it Works
